updated model
